% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate_int_vector.R
\name{disaggregate_int_vector}
\alias{disaggregate_int_vector}
\title{disaggregate_int_vector}
\usage{
disaggregate_int_vector(x, r, method)
}
\arguments{
\item{x}{An integer vector.}

\item{r}{A vector of proportions, sum to 1.}

\item{method}{The method used to distribute population. "random" (the default) put people randomly based on probabilities where higher decimals with higher probabilities. "rank" put people based on the decimal values from the highest to the lowest.}
}
\value{
An integer matrix
}
\description{
Disaggregate an integer vector into a integer matrix.
Row sums equals to the input integer vector.
Col sums equals to the input proportions multiplying the sum of the input vector.
}
\examples{
x <- sample.int(100,5)
r <- c(0.2, 0.8)
X \%o\% r
disaggregate_int_vector(x,r,method="random")

}
