% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_free_memory.R
\name{get_free_memory}
\alias{get_free_memory}
\title{get_free_memory}
\usage{
get_free_memory()
}
\value{
The free memory (MB).
}
\description{
A complete cross-platform function that checks the OS, retrieves the free memory
}
\examples{
free_mem <- get_free_memory()

}
