% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_grids.R
\name{int_grid}
\alias{int_grid}
\title{int_grid}
\usage{
int_grid(POP_path, admin_path, result_folder, workers)
}
\arguments{
\item{POP_path}{The file path of non-integer population estimates.}

\item{admin_path}{The file path of administrative units. Each cell with a unique admin unit id.}

\item{result_folder}{The folder to save results.}

\item{workers}{Optional. The count of available workers (CPU cores).}
}
\value{
The integer population raster will be saved in specified results folder.
}
\description{
Convert Floating Population Estimates into Integers
}
\examples{
POP_path <- "pop.tif" # the file path of population raster
pop_int <- int_grid(POP_path) # if admin units are unavailable
admin_path <- "admin.tif" # the file path of administrative units raster
pop_int <- int_grid(POP_path, admin_path)

}
