% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_groups.R
\name{int_groups}
\alias{int_groups}
\title{int_groups}
\usage{
int_groups(POP_path, group_path, admin_path, result_folder, method, workers)
}
\arguments{
\item{POP_path}{The file path of non-integer population estimates.}

\item{group_path}{The file path of proportions for each group. If there are multiple admin units, the variable name must be set as 'id'.}

\item{admin_path}{The file path of administrative units. Each cell with a unique admin unit id.}

\item{result_folder}{The folder to save results.}

\item{method}{The method used to distribute population. "random" (the default) put people based on the decimal value and input population count.}

\item{workers}{Optional. The count of available workers (CPU cores).}
}
\value{
A list of SpatRaster each with a single layer corresponding to each group.
}
\description{
Generate corresponding integer population count on each grid cell
for each group based on a vector of proportions for the interest groups
}
\examples{
POP_path <- "pop.tif" # the file path of population raster
admin_path <- "admin.tif" # the file path of administrative units raster
group_path <- "group.csv" #the file path of proportions of each group 
example of group data: group <- data.frame(id = 1:2,
                                           group_1 = runif(10),
                                           group_2 = 1 - group_1)
pop_int <- int_groups(POP_path, admin_path, group_path)

}
