% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_vector.R
\name{int_vector}
\alias{int_vector}
\title{int_vector}
\usage{
int_vector(x)
}
\arguments{
\item{x}{A numerical vector.}
}
\value{
An integer vector.
}
\description{
Generate corresponding integer vector with smallest difference to the input vector
}
\examples{
x <- runif(5) * 10
int_vector(x)

}
